==> Subreddit_Estimation/1. reddit-comments.ipynb <==
# Scrapping Entries from Reddit

There are two TV series that I absolutely love. They reflect the real life problems. That's why I scrapped the subreddits of Black Mirror and WestWorld. In this notebook, I will go through the scraping.

# libraries needed:
import requests
import time
import pandas as pd
import numpy as np

## Black Mirror Posts

To scrap the data, I used the api push. That enabled me to scrap data in less than a minute.

def get_author_comments(**kwargs):
    r = requests.get("https://api.pushshift.io/reddit/submission/search/",params=kwargs)
    data = r.json()
    return data['data']

before = None
all_posts = []
for i in range(20):
    print('grabbing {} posts...'.format((i+1)*500))
    posts = get_author_comments(subreddit="blackmirror",size=500,before=before,sort='desc',sort_type='created_utc')

    before = posts[-1]['created_utc'] # This will keep track of your position for the next call in the while loop
     
    all_posts.extend(posts)

    time.sleep(1)

At the end I have the following amount of posts:

len(all_posts)

Then I needed to check the number of unique ID's. That is very important because a common problem is having duplicated posts since 10000 is a big number. Thanks to push api, I had all different.

ids = []
for i in range(0,len(all_posts)):
    ids.append(all_posts[i]['id'])
len(set(ids))

Below is all the data related to the first post:

all_posts[0]

Here there are many columns that need to be extracted. The reason is I won't include them to the analysis. In order to do that, I extracted each element as one and then appended those to a list. The important columns are:
 - created_utc
 -   id
 -   is_video
 -   num_comments
 -   score
 -   selftext
 -   spoiler
 -   subreddit
 -   title
 
Below I created an empty lists for each column. Then I appended each element to these lists. When there is no element, then I added an empty element so that the list still has 10000 elements.

bm_utc = []
bm_id = []
bm_num_comments = []
bm_score = []
bm_selftext = []
bm_spoiler = []
bm_subreddit = []
bm_title = []

##############################

for i in range(0,len(all_posts)):
    bm_utc.append(all_posts[i]['created_utc'] if all_posts[i]['created_utc'] != '' else 'NA')
    bm_id.append(all_posts[i]['id'] if all_posts[i]['id'] != '' else None)
    #bm_is_video.append(all_posts[i]['is_video'] if all_posts[i]['is_video'] != '' else None)
    bm_num_comments.append(all_posts[i]['num_comments'] if all_posts[i]['num_comments'] != '' else None)
    bm_score.append(all_posts[i]['score'] if all_posts[i]['score'] != '' else None)
    bm_spoiler.append(all_posts[i]['spoiler'] if all_posts[i]['spoiler'] != '' else None)
    bm_subreddit.append(all_posts[i]['subreddit'] if all_posts[i]['subreddit'] != '' else None)
    bm_title.append(all_posts[i]['title'] if all_posts[i]['title'] != '' else None)


But the 'selftext' column is a little bit different than the rest. Since the empty values might create a KeyError, I used an try-except phrase. This helped me to bypass the error in case I need to. 

bm_text = []
bm_is_video = []
for i in range(0,len(all_posts)):
    try : all_posts[i]['selftext']
    except KeyError : all_posts[i]['selftext'] = ''
    bm_text.append(all_posts[i]['selftext'])
for i in range(0,len(all_posts)):
    try : all_posts[i]['is_video']
    except KeyError : all_posts[i]['is_video'] = ''
    bm_is_video.append(all_posts[i]['is_video'])

Below I created a dataframe for the posts from the subreddit of Black Mirror!

black_mirror = pd.DataFrame(
    {'created_utc' : bm_utc,
     'id': bm_id,
     'is_video' : bm_is_video,
     'num_comments' : bm_num_comments,
     'score' : bm_score,
     'selftext' : bm_text,
     'spoiler' : bm_spoiler,
     'subreddit' : bm_subreddit,
     'title': bm_title
    }
)

counter = 0
for i in range(0,len(black_mirror)):
    if black_mirror.selftext[i] == '':
        counter += 1
    else:
        continue

For our analysis, it is very important to know the number of the empty posts, in other words posts without body. With the help if the above code we see that there are 4310 empty elements. This number might vary since we are always pulling new data.

counter

## WestWorld Posts

To do the same for our next favourite tv series, I combined all the code:

def get_author_comments(**kwargs):
    r = requests.get("https://api.pushshift.io/reddit/submission/search/",params=kwargs)
    data = r.json()
    return data['data']

before = None
all_posts = []
for i in range(20):
    print('grabbing {} posts...'.format((i+1)*500))
    posts = get_author_comments(subreddit='westworld', size=500,before=before,sort='desc',sort_type='created_utc')

    before = posts[-1]['created_utc'] # This will keep track of your position for the next call in the while loop

    all_posts.extend(posts)

    time.sleep(1)

ww_utc = []
ww_id = []
ww_num_comments = []
ww_score = []
ww_selftext = []
ww_spoiler = []
ww_subreddit = []
ww_title = []

    ##############################

for i in range(0,len(all_posts)):
    ww_utc.append(all_posts[i]['created_utc'] if all_posts[i]['created_utc'] != '' else 'NA')
    ww_id.append(all_posts[i]['id'] if all_posts[i]['id'] != '' else None)
    #ww_is_video.append(all_posts[i]['is_video'] if all_posts[i]['is_video'] != '' else None)
    ww_num_comments.append(all_posts[i]['num_comments'] if all_posts[i]['num_comments'] != '' else None)
    ww_score.append(all_posts[i]['score'] if all_posts[i]['score'] != '' else None)
    ww_spoiler.append(all_posts[i]['spoiler'] if all_posts[i]['spoiler'] != '' else None)
    ww_subreddit.append(all_posts[i]['subreddit'] if all_posts[i]['subreddit'] != '' else None)
    ww_title.append(all_posts[i]['title'] if all_posts[i]['title'] != '' else None)

ww_text = []
ww_is_video = []

for i in range(0,len(all_posts)):
    try : all_posts[i]['selftext']
    except KeyError : all_posts[i]['selftext'] = ''
    ww_text.append(all_posts[i]['selftext'])
    
for i in range(0,len(all_posts)):
    try : all_posts[i]['is_video']
    except KeyError : all_posts[i]['is_video'] = ''
    ww_is_video.append(all_posts[i]['is_video'])

west_world = pd.DataFrame(
    {'created_utc' : ww_utc,
     'id': ww_id,
     'is_video' : ww_is_video,
     'num_comments' : ww_num_comments,
     'score' : ww_score,
     'selftext' : ww_text,
     'spoiler' : ww_spoiler,
     'subreddit' : ww_subreddit,
     'title': ww_title
    }
)

    ######################################

counter = 0
for i in range(0,len(west_world)):
    if west_world.selftext[i] == '':
        counter += 1
    else:
        continue
print(counter)

## Convert them to csv files

Let's check both of the dataframes one last time before converting them to csv files:

black_mirror.head()

west_world.head()

black_mirror.to_csv('blackmirror.csv')
west_world.to_csv('westworld.csv')
