==> Subreddit_Estimation/2. Scraping_Analysis.ipynb <==
# Scraping Analysis

## EDA

# Necessary libraries:
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer
from bs4 import BeautifulSoup        
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import RegexpTokenizer
import regex as re
from nltk.stem.porter import PorterStemmer

!pip install regex

After necessary packages are downloaded, we upload the csv files for the series that I scraped in the previous notebook.

blackmirror = pd.read_csv('./blackmirror.csv')
westworld = pd.read_csv('./westworld.csv')

blackmirror.head()

westworld.head()

There are unnecassary 'Unnamed: 0' column that can deleted. Then the NAs should be dealt. In order to do that I excluded all the empty cells since we had more than enough even when I exclude them for my models to run

blackmirror.drop('Unnamed: 0', inplace = True, axis = 1)
westworld.drop('Unnamed: 0', inplace = True, axis = 1)

blackmirror.isnull().sum()

westworld.isnull().sum()

blackmirror = blackmirror[blackmirror.selftext.notnull()]

blackmirror.shape[0]

westworld = westworld[westworld.selftext.notnull()]

westworld.shape[0]

series = pd.concat([blackmirror, westworld], axis = 0)

series.head()

Then some columns need some changes:
- is_video is replaced with 1 and 0's so that there are only numeric values
- spoiler column revised in a same way.
- subreddit column is so revised that instead of Blackmirror, the cells were assigned 1 and instead of WestWorld they were assigned 0.


series['is_video'].value_counts()

series.drop('is_video', inplace = True, axis = 1)

series['spoiler'].value_counts()

series['spoiler'] = series['spoiler'].astype(int)

series.subreddit = series.subreddit.map(lambda cell: 1 if cell == 'blackmirror' else 0)

series.head()

series.subreddit.value_counts(normalize = True)

series = (series[series.selftext != "[removed]"])
series = (series[series.selftext != "[deleted]"])

series = series.reset_index(drop=True)

series.head()

## Analysis

In the analysis section, I created a function with the name of 'prep_post'. That enables me to clear the html errors, remove the punctuations and english stop words and finally to stem the words.

### Preparation Function

def prep_post(raw_post, column):
    # Function to convert a raw review to a string of words
    # The input is a single string (a raw movie review), and 
    # the output is a single string (a preprocessed movie review)
    #
    for i in range(0,len(raw_post[column])):
        # 1. Remove HTML
        review_text = BeautifulSoup(raw_post[column][i]).get_text()
            #
            # 2. Remove non-letters        
        letters_only = re.sub("[^a-zA-Z]", " ", review_text)
           #
            # 3. Convert to lower case, split into individual words
        words = letters_only.lower().split()
            #
            # 4. In Python, searching a set is much faster than searching
            #   a list, so convert the stop words to a set
        stops = set(stopwords.words('english'))
            # 
            # 5. Remove stop words
        meaningful_words = [w for w in words if not w in stops]
            #
            # 6. Stemming
        p_stemmer = PorterStemmer()
        stem_post = [p_stemmer.stem(i) for i in meaningful_words]
            #
            # 7. Join the words back into one string separated by space, 
            # and return the result.
        raw_post[column][i] = (" ".join(stem_post))
    

This function is run on two columns:
- selftext
- title

The reason is that I want to see the prediction capabilities of the models based on text and the titles.

Finally the series dataframe is extracted with a csv file called 'series.csv'

prep_post(series, 'selftext')
prep_post(series, 'title')
# prep_post(X_test, 'selftext')
# prep_post(X_test, 'title')

series.to_csv('series.csv')
