==> Subreddit_Estimation/3. Modelling.ipynb <==
# Modelling

In the modelling notebook, I downloaded all the necessary libraries. Then created the training and testing data sets. These were then divided into further classes:
- training and testing for the reddit posts
- training and testing for the reddit post titles

import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
import numpy as np
from sklearn.metrics import accuracy_score
from sklearn.model_selection import cross_val_score, train_test_split, GridSearchCV
from sklearn.ensemble import ExtraTreesClassifier, RandomForestClassifier

series = pd.read_csv('./series.csv')

series.drop('Unnamed: 0',axis = 1, inplace = True)

series.dropna(inplace=True)

series.isnull().sum()

X_train, X_test, y_train, y_test = train_test_split(series[['id', 'selftext','title']],
                                                    series['subreddit'],
                                                    test_size = 0.25,
                                                    random_state = 42)

X_train = X_train.reset_index(drop = True)
y_train = y_train.reset_index(drop = True)
X_test = X_test.reset_index(drop = True)
y_test = y_test.reset_index(drop = True)

X_train.head()

## TF-IDF

In our case, using TF-IDF is a more accurate way to create the models because utilizing the frequencies is a more valid approach than using only the count of words in a certain observation.

There are two different models that I used. One is logistic regression and the other one is the random forest. I will build four models in total. Two for post body and two for titles. Let's see which one can be used to predict the series better:

### Selftext

vect = TfidfVectorizer()
X_text_train = vect.fit_transform(X_train['selftext'].tolist())
X_text_test = vect.transform(X_test['selftext'].tolist())
posts_train = pd.DataFrame(X_text_train.toarray(), columns=vect.get_feature_names())
posts_train.head(5)

### Now we can start modelling using the arrays!

#### Logistic Regression

lr = LogisticRegression()
lr.fit(X_text_train, y_train)

# Score on training data.
lr.score(X_text_train, y_train)

# Score on testing data.
lr.score(X_text_test, y_test)

X_text_test

pred_text = lr.predict(X_text_test)

accuracy_score(y_test, pred_text)

#### Random Forest

rf= RandomForestClassifier()
rf_param = {
    'n_estimators':[13,14,15,16,17],
    'max_depth':[11,12,13],
    'max_features':['auto',1.0]
}
gs = GridSearchCV(rf, param_grid=rf_param, cv=5)
gs.fit(X_text_train, y_train)
print(gs.best_score_)
gs.best_params_

### Title

vect = TfidfVectorizer()
X_title_train = vect.fit_transform(X_train['title'].tolist())
X_title_test = vect.transform(X_test['title'].tolist())
titles_train = pd.DataFrame(X_title_train.toarray(), columns=vect.get_feature_names())
titles_train.head(5)

### Now we can start modelling using the arrays!

#### Logistic Regression

lr = LogisticRegression()
lr.fit(X_title_train, y_train)

# Score on training data.
lr.score(X_title_train, y_train)

# Score on testing data.
lr.score(X_title_test, y_test)

pred_title = lr.predict(X_title_test)

accuracy_score(y_test, pred_title)

#### Random Forest

rf= RandomForestClassifier()
rf_param = {
    'n_estimators':[13,14,15,16,17],
    'max_depth':[15,16,17],
    'max_features':['auto',1.0]
}
gs = GridSearchCV(rf, param_grid=rf_param, cv=5)
gs.fit(X_title_train, y_train)
print(gs.best_score_)
gs.best_params_

## Conclusion

As conclusion, I can say that using logistic regression with the post bodies has a better predictive power than any other with the score of 92%
